data_integration/requirements.txt
# Project dependencies
# Run: pip install -r requirements.txt
Flask==2.0.1
SQLAlchemy==1.4.23
elasticsearch==7.17.0
descope==1.3.0
redis==4.0.2
celery==5.2.3
psycopg2-binary==2.9.3
python-dotenv==0.19.2
